+++ airgeddon.sh
@@ -2,8 +2,8 @@
 #Title........: airgeddon.sh
 #Description..: This is a multi-use bash script for Linux systems to audit wireless networks.
 #Author.......: v1s1t0r
-#Date.........: 20170525
-#Version......: 7.1
+#Date.........: 20170614
+#Version......: 7.11
 #Usage........: bash airgeddon.sh
 #Bash Version.: 4.2 or later
 
@@ -15,6 +15,9 @@ auto_update=0
 
 #Enabled 1 / Disabled 0 - Auto change language feature - Default value 1
 auto_change_language=1
+
+#Enabled 1 / Disabled 0 - Allow colorized output - Default value 1
+allow_colorization=1
 
 #Language vars
 #Change this line to select another default language. Select one from available values in array
@@ -106,13 +109,12 @@ declare -A possible_alias_names=(
 								)
 
 #General vars
-airgeddon_version="7.1"
-language_strings_expected_version="7.1-1"
+airgeddon_version="7.11"
+language_strings_expected_version="7.11-1"
 standardhandshake_filename="handshake-01.cap"
 tmpdir="/tmp/"
 osversionfile_dir="/etc/"
 minimum_bash_version_required="4.2"
-hashcat3_version="3.0"
 resume_message=224
 abort_question=12
 pending_of_translation="[PoT]"
@@ -121,6 +123,22 @@ standard_resolution="1024x768"
 curl_404_error="404: Not Found"
 language_strings_file="language_strings.sh"
 broadcast_mac="FF:FF:FF:FF:FF:FF"
+
+#aircrack vars
+aircrack_tmp_simple_name_file="aircrack"
+aircrack_pot_tmp="${aircrack_tmp_simple_name_file}.pot"
+
+#hashcat vars
+hashcat3_version="3.0"
+hashcat_hccapx_version="3.40"
+hashcat_tmp_simple_name_file="hctmp"
+hashcat_tmp_file="${hashcat_tmp_simple_name_file}.hccap"
+hashcat_pot_tmp="${hashcat_tmp_simple_name_file}.pot"
+hashcat_output_file="${hashcat_tmp_simple_name_file}.out"
+hccapx_tool="cap2hccapx"
+possible_hccapx_converter_known_locations=(
+										"/usr/lib/hashcat-utils/${hccapx_tool}.bin"
+									)
 
 #WEP vars
 wep_data="wepdata"
@@ -157,6 +175,7 @@ urlscript_pins_dbfile="https://raw.githubusercontent.com/${github_user}/${github
 urlscript_pins_dbfile_checksum="https://raw.githubusercontent.com/${github_user}/${github_repository}/${branch}/${pins_dbfile_checksum}"
 urlscript_language_strings_file="https://raw.githubusercontent.com/${github_user}/${github_repository}/${branch}/${language_strings_file}"
 urlgithub_wiki="https://${repository_hostname}/${github_user}/${github_repository}/wiki"
+bitcoin="1AKnTXbomtwUzrm81FRzi5acSSXxGteGTH"
 mail="v1s1t0r.1s.h3r3@gmail.com"
 author="v1s1t0r"
 
@@ -249,7 +268,7 @@ known_arm_compatible_distros=(
 							)
 
 #Hint vars
-declare main_hints=(128 134 163)
+declare main_hints=(128 134 163 437 438 442)
 declare dos_hints=(129 131 133)
 declare handshake_hints=(127 130 132 136)
 declare handshake_attack_hints=(142)
@@ -720,6 +739,19 @@ function check_interface_coherence() {
 	fi
 
 	return ${interface_auto_change}
+}
+
+#Add contributing footer to a file
+function add_contributing_footer_to_file() {
+
+	debug_print
+
+	{
+	echo ""
+	echo "---------------"
+	echo ""
+	echo "${footer_texts[${language},1]}"
+	} >> "${1}"
 }
 
 #Prepare the vars to be used on wps pin database attacks
@@ -1476,18 +1508,33 @@ function ask_bssid() {
 	local regexp="^([a-fA-F0-9]{2}:){5}[a-zA-Z0-9]{2}$"
 
 	if [ "${1}" = "wps" ]; then
+		if [ -z "${wps_bssid}" ]; then
+			ask_yesno 439 "no"
+			if [ ${yesno} = "n" ]; then
+				return 1
+			fi
+		fi
+
 		while [[ ! ${wps_bssid} =~ ${regexp} ]]; do
 			read_bssid "wps"
 		done
 		echo
 		language_strings "${language}" 364 "blue"
 	else
+		if [ -z "${bssid}" ]; then
+			ask_yesno 439 "no"
+			if [ ${yesno} = "n" ]; then
+				return 1
+			fi
+		fi
+
 		while [[ ! ${bssid} =~ ${regexp} ]]; do
 			read_bssid
 		done
 		echo
 		language_strings "${language}" 28 "blue"
 	fi
+	return 0
 }
 
 #Read the user input on essid questions
@@ -1506,6 +1553,14 @@ function ask_essid() {
 	debug_print
 
 	if [ -z "${essid}" ]; then
+
+		if [ "${1}" = "verify" ]; then
+			ask_yesno 439 "no"
+			if [ ${yesno} = "n" ]; then
+				return 1
+			fi
+		fi
+
 		while [[ -z "${essid}" ]]; do
 			read_essid
 		done
@@ -1742,9 +1797,17 @@ function set_wep_key_script() {
 
 	cat >&8 <<-'EOF'
 				echo -en " ${wep_hex_key}"
+				echo ""
 	EOF
 
 	cat >&8 <<-EOF
+				} >> "${weppotenteredpath}"
+				
+				{
+				echo ""
+				echo "---------------"
+				echo ""
+				echo "${footer_texts[${language},1]}"
 				} >> "${weppotenteredpath}"
 			fi
 		}
@@ -2379,6 +2442,9 @@ function mdk3_deauth_option() {
 	language_strings "${language}" 34 "yellow"
 
 	ask_bssid
+	if [ "$?" != "0" ]; then
+		return
+	fi
 	ask_channel
 	exec_mdk3deauth
 }
@@ -2401,6 +2467,9 @@ function aireplay_deauth_option() {
 	language_strings "${language}" 34 "yellow"
 
 	ask_bssid
+	if [ "$?" != "0" ]; then
+		return
+	fi
 	ask_channel
 	exec_aireplaydeauth
 }
@@ -2422,7 +2491,10 @@ function wds_confusion_option() {
 	echo
 	language_strings "${language}" 34 "yellow"
 
-	ask_essid
+	ask_essid "verify"
+	if [ "$?" != "0" ]; then
+		return
+	fi
 	ask_channel
 	exec_wdsconfusion
 }
@@ -2444,7 +2516,10 @@ function beacon_flood_option() {
 	echo
 	language_strings "${language}" 34 "yellow"
 
-	ask_essid
+	ask_essid "verify"
+	if [ "$?" != "0" ]; then
+		return
+	fi
 	ask_channel
 	exec_beaconflood
 }
@@ -2467,6 +2542,9 @@ function auth_dos_option() {
 	language_strings "${language}" 34 "yellow"
 
 	ask_bssid
+	if [ "$?" != "0" ]; then
+		return
+	fi
 	exec_authdos
 }
 
@@ -2488,6 +2566,9 @@ function michael_shutdown_option() {
 	language_strings "${language}" 34 "yellow"
 
 	ask_bssid
+	if [ "$?" != "0" ]; then
+		return
+	fi
 	exec_michaelshutdown
 }
 
@@ -2541,6 +2622,9 @@ function wps_attacks_parameters() {
 	language_strings "${language}" 34 "yellow"
 
 	ask_bssid "wps"
+	if [ "$?" != "0" ]; then
+		return 1
+	fi
 	ask_channel "wps"
 
 	case ${wps_attack} in
@@ -2881,6 +2965,7 @@ function clean_tmpfiles() {
 	rm -rf "${tmpdir}clts"* > /dev/null 2>&1
 	rm -rf "${tmpdir}wnws.txt" > /dev/null 2>&1
 	rm -rf "${tmpdir}hctmp"* > /dev/null 2>&1
+	rm -rf "${tmpdir}${aircrack_pot_tmp}" > /dev/null 2>&1
 	rm -rf "${tmpdir}${hostapd_file}" > /dev/null 2>&1
 	rm -rf "${tmpdir}${dhcpd_file}" > /dev/null 2>&1
 	rm -rf "${tmpdir}${control_file}" > /dev/null 2>&1
@@ -3971,6 +4056,7 @@ function aircrack_dictionary_attack_option() {
 	language_strings "${language}" 190 "yellow"
 	language_strings "${language}" 115 "read"
 	exec_aircrack_dictionary_attack
+	manage_aircrack_pot
 }
 
 #Validate and ask for the different parameters used in an aircrack bruteforce based attack
@@ -3998,6 +4084,7 @@ function aircrack_bruteforce_attack_option() {
 	language_strings "${language}" 190 "yellow"
 	language_strings "${language}" 115 "read"
 	exec_aircrack_bruteforce_attack
+	manage_aircrack_pot
 }
 
 #Validate and ask for the different parameters used in a hashcat dictionary based attack
@@ -4008,6 +4095,11 @@ function hashcat_dictionary_attack_option() {
 	manage_asking_for_captured_file
 
 	select_wpa_bssid_target_from_captured_file "${enteredpath}"
+	if [ "$?" != "0" ]; then
+		return
+	fi
+
+	convert_cap_to_hashcat_format
 	if [ "$?" != "0" ]; then
 		return
 	fi
@@ -4029,6 +4121,11 @@ function hashcat_bruteforce_attack_option() {
 	manage_asking_for_captured_file
 
 	select_wpa_bssid_target_from_captured_file "${enteredpath}"
+	if [ "$?" != "0" ]; then
+		return
+	fi
+
+	convert_cap_to_hashcat_format
 	if [ "$?" != "0" ]; then
 		return
 	fi
@@ -4061,8 +4158,12 @@ function hashcat_rulebased_attack_option() {
 		return
 	fi
 
-	manage_asking_for_dictionary_file
+	convert_cap_to_hashcat_format
+	if [ "$?" != "0" ]; then
+		return
+	fi
 
+	manage_asking_for_dictionary_file
 	manage_asking_for_rule_file
 
 	echo
@@ -4077,12 +4178,28 @@ function manage_hashcat_pot() {
 
 	debug_print
 
-	local regexp="All hashes have been recovered"
-	if [ -n "${hashcat_fix}" ]; then
-		local regexp="Status\.{1,9}:[[:space:]]Cracked"
+	hashcat_output=$(cat "${tmpdir}${hashcat_output_file}")
+
+	pass_decrypted_by_hashcat=0
+	if compare_floats_greater_or_equal "${hashcat_version}" "${hashcat3_version}"; then
+		local regexp="Status\.+:[[:space:]]Cracked"
+		if [[ ${hashcat_output} =~ ${regexp} ]]; then
+			pass_decrypted_by_hashcat=1
+		else
+			if compare_floats_greater_or_equal "${hashcat_version}" "${hashcat_hccapx_version}"; then
+				if [[ -f "${tmpdir}${hashcat_pot_tmp}" ]]; then
+					pass_decrypted_by_hashcat=1
+				fi
+			fi
+		fi
+	else
+		local regexp="All hashes have been recovered"
+		if [[ ${hashcat_output} =~ ${regexp} ]]; then
+			pass_decrypted_by_hashcat=1
+		fi
 	fi
 
-	if [[ ${hashcat_output} =~ ${regexp} ]]; then
+	if [ "${pass_decrypted_by_hashcat}" -eq 1 ]; then
 
 		echo
 		language_strings "${language}" 234 "yellow"
@@ -4094,7 +4211,7 @@ function manage_hashcat_pot() {
 			if [ "${lastcharhashcat_potpath}" != "/" ]; then
 				hashcat_potpath="${hashcat_potpath}/"
 			fi
-			hashcatpot_filename="hashcat-${bssid}.pot"
+			hashcatpot_filename="hashcat-${bssid}.txt"
 			hashcat_potpath="${hashcat_potpath}${hashcatpot_filename}"
 
 			validpath=1
@@ -4102,9 +4219,74 @@ function manage_hashcat_pot() {
 				read_path "hashcatpot"
 			done
 
-			cp "${tmpdir}hctmp.pot" "${potenteredpath}"
+			[[ $(cat "${tmpdir}${hashcat_pot_tmp}") =~ .+:(.+)$ ]] && hashcat_key="${BASH_REMATCH[1]}"
+			{
+			echo ""
+			date +%Y-%m-%d
+			echo "${hashcat_texts[${language},1]}"
+			echo ""
+			echo "BSSID: ${bssid}"
+			echo ""
+			echo "---------------"
+			echo ""
+			echo "${hashcat_key}"
+			} >> "${potenteredpath}"
+
+			add_contributing_footer_to_file "${potenteredpath}"
+
 			echo
 			language_strings "${language}" 236 "blue"
+			language_strings "${language}" 115 "read"
+		fi
+	fi
+}
+
+#Check if the password was decrypted using aircrack and manage to save it on a file
+function manage_aircrack_pot() {
+
+	debug_print
+
+	pass_decrypted_by_aircrack=0
+	if [ -f "${tmpdir}${aircrack_pot_tmp}" ]; then
+		pass_decrypted_by_aircrack=1
+	fi
+
+	if [ "${pass_decrypted_by_aircrack}" -eq 1 ]; then
+
+		echo
+		language_strings "${language}" 234 "yellow"
+		ask_yesno 235 "yes"
+		if [ ${yesno} = "y" ]; then
+			aircrack_potpath="${default_save_path}"
+			lastcharaircrack_potpath=${aircrack_potpath: -1}
+			if [ "${lastcharaircrack_potpath}" != "/" ]; then
+				aircrack_potpath="${aircrack_potpath}/"
+			fi
+			aircrackpot_filename="aircrack-${bssid}.txt"
+			aircrack_potpath="${aircrack_potpath}${aircrackpot_filename}"
+
+			validpath=1
+			while [[ "${validpath}" != "0" ]]; do
+				read_path "aircrackpot"
+			done
+
+			aircrack_key=$(cat "${tmpdir}${aircrack_pot_tmp}")
+			{
+			echo ""
+			date +%Y-%m-%d
+			echo "${aircrack_texts[${language},1]}"
+			echo ""
+			echo "BSSID: ${bssid}"
+			echo ""
+			echo "---------------"
+			echo ""
+			echo "${aircrack_key}"
+			} >> "${aircrackpotenteredpath}"
+
+			add_contributing_footer_to_file "${aircrackpotenteredpath}"
+
+			echo
+			language_strings "${language}" 440 "blue"
 			language_strings "${language}" 115 "read"
 		fi
 	fi
@@ -4418,8 +4600,9 @@ function set_show_charset() {
 function exec_aircrack_bruteforce_attack() {
 
 	debug_print
-
-	crunch "${minlength}" "${maxlength}" "${charset}" | aircrack-ng -a 2 -b "${bssid}" -w - "${enteredpath}"
+	rm -rf "${tmpdir}${aircrack_pot_tmp}" > /dev/null 2>&1
+	aircrack_cmd="crunch \"${minlength}\" \"${maxlength}\" \"${charset}\" | aircrack-ng -a 2 -b \"${bssid}\" -l \"${tmpdir}${aircrack_pot_tmp}\" -w - \"${enteredpath}\" ${colorize}"
+	eval "${aircrack_cmd}"
 	language_strings "${language}" 115 "read"
 }
 
@@ -4428,7 +4611,9 @@ function exec_aircrack_dictionary_attack() {
 
 	debug_print
 
-	aircrack-ng -a 2 -b "${bssid}" -w "${DICTIONARY}" "${enteredpath}"
+	rm -rf "${tmpdir}${aircrack_pot_tmp}" > /dev/null 2>&1
+	aircrack_cmd="aircrack-ng -a 2 -b \"${bssid}\" -l \"${tmpdir}${aircrack_pot_tmp}\" -w \"${DICTIONARY}\" \"${enteredpath}\" ${colorize}"
+	eval "${aircrack_cmd}"
 	language_strings "${language}" 115 "read"
 }
 
@@ -4437,10 +4622,8 @@ function exec_hashcat_dictionary_attack() {
 
 	debug_print
 
-	convert_cap_to_hashcat_format
-	hashcat_cmd="hashcat -m 2500 -a 0 \"${tmpdir}hctmp.hccap\" \"${DICTIONARY}\" --potfile-disable -o \"${tmpdir}hctmp.pot\" ${hashcat_fix} | tee /dev/fd/5"
-	exec 5>&1
-	hashcat_output=$(eval "${hashcat_cmd}")
+	hashcat_cmd="hashcat -m 2500 -a 0 \"${tmpdir}${hashcat_tmp_file}\" \"${DICTIONARY}\" --potfile-disable -o \"${tmpdir}${hashcat_pot_tmp}\"${hashcat_fix} | tee \"${tmpdir}${hashcat_output_file}\" ${colorize}"
+	eval "${hashcat_cmd}"
 	language_strings "${language}" 115 "read"
 }
 
@@ -4449,10 +4632,8 @@ function exec_hashcat_bruteforce_attack() {
 
 	debug_print
 
-	convert_cap_to_hashcat_format
-	hashcat_cmd="hashcat -m 2500 -a 3 \"${tmpdir}hctmp.hccap\" \"${charset}\" --potfile-disable -o \"${tmpdir}hctmp.pot\" ${hashcat_fix} | tee /dev/fd/5"
-	exec 5>&1
-	hashcat_output=$(eval "${hashcat_cmd}")
+	hashcat_cmd="hashcat -m 2500 -a 3 \"${tmpdir}${hashcat_tmp_file}\" \"${charset}\" --potfile-disable -o \"${tmpdir}${hashcat_pot_tmp}\"${hashcat_fix} | tee \"${tmpdir}${hashcat_output_file}\" ${colorize}"
+	eval "${hashcat_cmd}"
 	language_strings "${language}" 115 "read"
 }
 
@@ -4461,10 +4642,8 @@ function exec_hashcat_rulebased_attack() {
 
 	debug_print
 
-	convert_cap_to_hashcat_format
-	hashcat_cmd="hashcat -m 2500 -a 0 \"${tmpdir}hctmp.hccap\" \"${DICTIONARY}\" -r \"${RULES}\" --potfile-disable -o \"${tmpdir}hctmp.pot\" ${hashcat_fix} | tee /dev/fd/5"
-	exec 5>&1
-	hashcat_output=$(eval "${hashcat_cmd}")
+	hashcat_cmd="hashcat -m 2500 -a 0 \"${tmpdir}${hashcat_tmp_file}\" \"${DICTIONARY}\" -r \"${RULES}\" --potfile-disable -o \"${tmpdir}${hashcat_pot_tmp}\"${hashcat_fix} | tee \"${tmpdir}${hashcat_output_file}\" ${colorize}"
+	eval "${hashcat_cmd}"
 	language_strings "${language}" 115 "read"
 }
 
@@ -4949,6 +5128,7 @@ function set_wps_attack_script() {
 		script_interface="${interface}"
 		script_wps_bssid="${wps_bssid}"
 		script_wps_channel="${wps_channel}"
+		colorize="${colorize}"
 	EOF
 
 	cat >&7 <<-'EOF'
@@ -5137,7 +5317,7 @@ function set_wps_attack_script() {
 					fi
 
 					this_pin_timeout=0
-					(set -o pipefail && eval "${script_attack_cmd1}${current_pin}${script_attack_cmd2}")
+					(set -o pipefail && eval "${script_attack_cmd1}${current_pin}${script_attack_cmd2} ${colorize}")
 					if [ "$?" = "124" ]; then
 						if [ "${script_wps_attack_tool}" = "reaver" ]; then
 							this_pin_timeout=1
@@ -5179,7 +5359,7 @@ function set_wps_attack_script() {
 					echo
 				fi
 
-				(set -o pipefail && eval "${script_attack_cmd1}${current_pin}${script_attack_cmd2}")
+				(set -o pipefail && eval "${script_attack_cmd1}${current_pin}${script_attack_cmd2} ${colorize}")
 				if [ "$?" = "124" ]; then
 					if [ "${script_wps_attack_tool}" = "reaver" ]; then
 						this_pin_timeout=1
@@ -5209,7 +5389,7 @@ function set_wps_attack_script() {
 					echo
 				fi
 
-				(set -o pipefail && eval "${script_attack_cmd1}${script_attack_cmd2}")
+				(set -o pipefail && eval "${script_attack_cmd1}${script_attack_cmd2} ${colorize}")
 				if [ "$?" = "124" ]; then
 					this_pin_timeout=1
 				fi
@@ -5221,7 +5401,7 @@ function set_wps_attack_script() {
 				if [ "${script_wps_attack_tool}" = "bully" ]; then
 					echo
 				fi
-				eval "${script_attack_cmd1}${script_attack_cmd2}"
+				eval "${script_attack_cmd1}${script_attack_cmd2} ${colorize}"
 				parse_output
 			;;
 		esac
@@ -5357,6 +5537,13 @@ function set_control_script() {
 						"${et_captive_portal_logpath}"
 					done
 				fi
+				
+				{
+				echo ""
+				echo "---------------"
+				echo ""
+				echo "${footer_texts[${language},1]}"
+				} >> "${et_captive_portal_logpath}"
 
 				sleep 2
 				killall hostapd > /dev/null 2>&1
@@ -6230,7 +6417,35 @@ function convert_cap_to_hashcat_format() {
 
 	tmpfiles_toclean=1
 	rm -rf "${tmpdir}hctmp"* > /dev/null 2>&1
-	echo "1" | aircrack-ng "${enteredpath}" -J "${tmpdir}hctmp" -b "${bssid}" > /dev/null 2>&1
+	if [ "${hccapx_needed}" -eq 0 ]; then
+		echo "1" | aircrack-ng "${enteredpath}" -J "${tmpdir}${hashcat_tmp_simple_name_file}" -b "${bssid}" > /dev/null 2>&1
+		return 0
+	else
+		hccapx_converter_found=0
+		if hash ${hccapx_tool} 2> /dev/null; then
+			hccapx_converter_found=1
+			hccapx_converter_path="${hccapx_tool}"
+		else
+			for item in "${possible_hccapx_converter_known_locations[@]}"; do
+				if [ -f "${item}" ]; then
+					hccapx_converter_found=1
+					hccapx_converter_path="${item}"
+					break
+				fi
+			done
+		fi
+
+		if [ "${hccapx_converter_found}" -eq 1 ]; then
+			hashcat_tmp_file="${hashcat_tmp_simple_name_file}.hccapx"
+			"${hccapx_converter_path}" "${enteredpath}" "${tmpdir}${hashcat_tmp_file}" > /dev/null 2>&1
+			return 0
+		else
+			echo
+			language_strings "${language}" 436 "red"
+			language_strings "${language}" 115 "read"
+			return 1
+		fi
+	fi
 }
 
 #Handshake tools menu
@@ -6585,6 +6800,10 @@ function validate_path() {
 				enteredpath="${pathname}${standardhandshake_filename}"
 				suggested_filename="${standardhandshake_filename}"
 			;;
+			"aircrackpot")
+				suggested_filename="${aircrackpot_filename}"
+				aircrackpotenteredpath+="${aircrackpot_filename}"
+			;;
 			"hashcatpot")
 				suggested_filename="${hashcatpot_filename}"
 				potenteredpath+="${hashcatpot_filename}"
@@ -6677,6 +6896,14 @@ function read_path() {
 			language_strings "${language}" 242 "green"
 			read_and_clean_path "RULES"
 			check_file_exists "${RULES}"
+		;;
+		"aircrackpot")
+			language_strings "${language}" 441 "green"
+			read_and_clean_path "aircrackpotenteredpath"
+			if [ -z "${aircrackpotenteredpath}" ]; then
+				aircrackpotenteredpath="${aircrack_potpath}"
+			fi
+			validate_path "${aircrackpotenteredpath}" "${1}"
 		;;
 		"hashcatpot")
 			language_strings "${language}" 233 "green"
@@ -7358,8 +7585,12 @@ function et_prerequisites() {
 		language_strings "${language}" 31 "blue"
 	else
 		ask_bssid
+		if [ "$?" != "0" ]; then
+			return_to_et_main_menu=1
+			return
+		fi
 		ask_channel
-		ask_essid
+		ask_essid "noverify"
 	fi
 
 	if [[ "${et_mode}" = "et_sniffing" ]] || [[ "${et_mode}" = "et_sniffing_sslstrip" ]]; then
@@ -7942,8 +8173,11 @@ function set_hashcat_parameters() {
 	hashcat_fix=""
 	hashcat_charset_fix_needed=0
 	if compare_floats_greater_or_equal "${hashcat_version}" "${hashcat3_version}"; then
-		hashcat_fix=" -D 1 --force"
+		hashcat_fix=" --weak-hash-threshold 0 -D 1 --force"
 		hashcat_charset_fix_needed=1
+		if compare_floats_greater_or_equal "${hashcat_version}" "${hashcat_hccapx_version}"; then
+			hccapx_needed=1
+		fi
 	fi
 }
 
@@ -8782,6 +9016,7 @@ function initialize_script_settings() {
 	fake_beef_found=0
 	set_script_folder_and_name
 	http_proxy_set=0
+	hccapx_needed=0
 }
 
 #Detect screen resolution if possible
@@ -8922,6 +9157,19 @@ function docker_detection() {
 	fi
 }
 
+#Set colorization output if set
+function initialize_colorized_output() {
+
+	debug_print
+
+	colorize=""
+	if [ "${allow_colorization}" -eq 1 ]; then
+		if hash ccze 2> /dev/null; then
+			colorize="| ccze -A"
+		fi
+	fi
+}
+
 #Script starting point
 function welcome() {
 
@@ -8990,6 +9238,7 @@ function welcome() {
 		check_update_tools
 	fi
 
+	initialize_colorized_output
 	set_windows_sizes
 	select_interface
 	initialize_menu_options_dependencies